auth

+++ frontend/modules/direktor/controllers/OcenaController.php
@@ -5,10 +5,11 @@ namespace frontend\modules\direktor\controllers;
 use Yii;
 use frontend\modules\direktor\models\Ocena;
 use frontend\modules\direktor\models\OcenaSearch;
+use yii\filters\AccessControl;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
-
+use yii\web\ForbiddenHttpException;
 use frontend\modules\direktor\models\Predmet;
 use frontend\modules\direktor\models\Odeljenje;
 
@@ -38,15 +39,23 @@ class OcenaController extends Controller
      */
     public function actionIndex()
     {
-        $searchModel = new OcenaSearch();
-        $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
-        $predmet = Predmet::find()->select('id_predmet,naziv')->all();
+            if(Yii::$app->user->can('direktor')){
+                $searchModel = new OcenaSearch();
+                $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
+                $predmet = Predmet::find()->select('id_predmet,naziv')->all();
 
-        return $this->render('index', [
-            'predmet' => $predmet,
-            'searchModel' => $searchModel,
-            'dataProvider' => $dataProvider,
-        ]);
+                return $this->render('index', [
+                    'predmet' => $predmet,
+                    'searchModel' => $searchModel,
+                    'dataProvider' => $dataProvider,
+                ]);
+            } else if(Yii::$app->user->isGuest){
+                $this->redirect(['../site/login']);
+            } else {
+                throw new ForbiddenHttpException('Nemate pravo pristupa ovoj stranici');
+            }
+
+
     }
 
     /**
@@ -57,17 +66,22 @@ class OcenaController extends Controller
      */
     public function actionView($id)
     {
+        if(Yii::$app->user->can('direktor')){
         $predmet = Predmet::find()->select('id_predmet,naziv')->where(['id_predmet' => $id])->all();
         $id_predmet = $id;
         $q = Yii::$app->db->createCommand("SELECT AVG(`ocena`.`zakljucena_ocena`) AS Prosek 
                                                   FROM `ocena` WHERE `ocena`.id_predmet = $id_predmet");
         $model= $q->queryAll();
 
-
         return $this->render('view', [
             'model' => $model,
             'predmet' => $predmet,
         ]);
+        } else if(Yii::$app->user->isGuest){
+            $this->redirect(['../site/login']);
+        } else {
+            throw new ForbiddenHttpException('Nemate pravo pristupa ovoj stranici');
+        }
     }
 
     /**
@@ -77,6 +91,7 @@ class OcenaController extends Controller
      */
     public function actionDirektor()
     {
+        if(Yii::$app->user->can('direktor')){
         $searchModel = new OcenaSearch();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
         $predmet = Predmet::find()->select('id_predmet,naziv')->all();
@@ -88,6 +103,11 @@ class OcenaController extends Controller
             'searchModel' => $searchModel,
             'dataProvider' => $dataProvider,
         ]);
+        } else if(Yii::$app->user->isGuest){
+            $this->redirect(['../site/login']);
+        } else {
+            throw new ForbiddenHttpException('Nemate pravo pristupa ovoj stranici');
+        }
     }
 
     /**
@@ -99,22 +119,29 @@ class OcenaController extends Controller
      */
     public function actionDirekto($id, $idp)
     {
-        $predmet = Predmet::find()->select('id_predmet,naziv')->where(['id_predmet' => $id])->all();
-        $id_predmet = $id;
-        $odeljenje = Odeljenje::find()->select('id_odeljenje,naziv')->where(['id_odeljenje' => $idp])->all();
-        $id_odeljenje = $idp;
-        $q = Yii::$app->db->createCommand("SELECT AVG(`ocena`.`zakljucena_ocena`) AS Prosek FROM `ocena`
+        if(Yii::$app->user->can('direktor')){
+            $predmet = Predmet::find()->select('id_predmet,naziv')->where(['id_predmet' => $id])->all();
+            $id_predmet = $id;
+            $odeljenje = Odeljenje::find()->select('id_odeljenje,naziv')->where(['id_odeljenje' => $idp])->all();
+            $id_odeljenje = $idp;
+            $q = Yii::$app->db->createCommand("SELECT AVG(`ocena`.`zakljucena_ocena`) AS Prosek FROM `ocena`
                                                     JOIN `ucenik` ON `ocena`.`id_ucenik`=`ucenik`.`id_ucenik`
                                                     JOIN `odeljenje` ON `ucenik`.`id_odeljenje`=`odeljenje`.`id_odeljenje`
                                                     WHERE `ocena`.`id_predmet`= $id_predmet AND `odeljenje`.`id_odeljenje`= $id_odeljenje");
-        $model= $q->queryAll();
+            $model= $q->queryAll();
+
+            return $this->render('dview', [
+                'model' => $model,
+                'predmet' => $predmet,
+                'odeljenje' => $odeljenje,
+            ]);
+        } else if(Yii::$app->user->isGuest){
+            $this->redirect(['../site/login']);
+        } else {
+            throw new ForbiddenHttpException('Nemate pravo pristupa ovoj stranici');
+        }
 
 
-        return $this->render('dview', [
-            'model' => $model,
-            'predmet' => $predmet,
-            'odeljenje' => $odeljenje,
-        ]);
     }
     
     /**
@@ -126,9 +153,17 @@ class OcenaController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        if(Yii::$app->user->can('direktor')){
+            $this->findModel($id)->delete();
 
-        return $this->redirect(['index']);
+            return $this->redirect(['index']);
+        } else if(Yii::$app->user->isGuest){
+            $this->redirect(['site/login']);
+        } else if(Yii::$app->user->isGuest){
+            $this->redirect(['../site/login']);
+        } else {
+            throw new ForbiddenHttpException('Nemate pravo pristupa ovoj stranici');
+        }
     }
 
     /**
